+++ crates/astria-account-monitor/src/metrics.rs
@@ -6,6 +6,7 @@ use telemetry::{
         Counter,
         Error,
         Gauge,
+        Recorder,
         RegisteringBuilder,
     },
 };
@@ -64,7 +65,10 @@ impl Metrics {
 impl telemetry::Metrics for Metrics {
     type Config = crate::Config;
 
-    fn register(builder: &mut RegisteringBuilder, config: &Self::Config) -> Result<Self, Error>
+    fn register<R: Recorder>(
+        builder: &mut RegisteringBuilder<R>,
+        config: &Self::Config,
+    ) -> Result<Self, Error>
     where
         Self: Sized,
     {