+++ crates/astria-sequencer-relayer/src/metrics.rs
@@ -6,6 +6,7 @@ use telemetry::{
         Counter,
         Gauge,
         Histogram,
+        Recorder,
         RegisteringBuilder,
     },
 };
@@ -98,8 +99,8 @@ impl telemetry::Metrics for Metrics {
         reason = "this is reasonable as we have a lot of metrics to register; the function is not \
                   complex, just long"
     )]
-    fn register(
-        builder: &mut RegisteringBuilder,
+    fn register<R: Recorder>(
+        builder: &mut RegisteringBuilder<R>,
         _config: &Self::Config,
     ) -> Result<Self, telemetry::metrics::Error> {
         let celestia_submission_height = builder