+++ crates/astria-telemetry/src/metrics/mod.rs
@@ -6,8 +6,9 @@ mod gauge;
 mod handle;
 mod histogram;
 mod into_f64;
+mod noop_recorder;
 
-use metrics_exporter_prometheus::PrometheusBuilder;
+pub use metrics::Recorder;
 
 pub use self::{
     builders::{
@@ -48,12 +49,15 @@ pub trait Metrics {
     /// # Errors
     ///
     /// Implementations should return an error if registering metrics fails.
-    fn register(builder: &mut RegisteringBuilder, config: &Self::Config) -> Result<Self, Error>
+    fn register<R: Recorder>(
+        builder: &mut RegisteringBuilder<R>,
+        config: &Self::Config,
+    ) -> Result<Self, Error>
     where
         Self: Sized;
 
-    /// Returns an instance of `Self` where the metrics are registered to a recorder that is
-    /// dropped immediately, meaning metrics aren't recorded.
+    /// Returns an instance of `Self` where the metrics are registered to a recorder that does
+    /// nothing, meaning metrics aren't recorded.
     ///
     /// # Errors
     ///
@@ -62,7 +66,7 @@ pub trait Metrics {
     where
         Self: Sized,
     {
-        let mut builder = RegisteringBuilder::new(PrometheusBuilder::new().build_recorder());
+        let mut builder = RegisteringBuilder::new(noop_recorder::NoopRecorder);
         Self::register(&mut builder, config)
     }
 }