+++ crates/astria-telemetry/src/metrics/noop_recorder.rs
@@ -1,0 +1,33 @@
+use metrics::{
+    Counter,
+    Gauge,
+    Histogram,
+    Key,
+    KeyName,
+    Metadata,
+    Recorder,
+    SharedString,
+    Unit,
+};
+
+pub(super) struct NoopRecorder;
+
+impl Recorder for NoopRecorder {
+    fn describe_counter(&self, _key: KeyName, _unit: Option<Unit>, _description: SharedString) {}
+
+    fn describe_gauge(&self, _key: KeyName, _unit: Option<Unit>, _description: SharedString) {}
+
+    fn describe_histogram(&self, _key: KeyName, _unit: Option<Unit>, _description: SharedString) {}
+
+    fn register_counter(&self, _key: &Key, _metadata: &Metadata<'_>) -> Counter {
+        Counter::noop()
+    }
+
+    fn register_gauge(&self, _key: &Key, _metadata: &Metadata<'_>) -> Gauge {
+        Gauge::noop()
+    }
+
+    fn register_histogram(&self, _key: &Key, _metadata: &Metadata<'_>) -> Histogram {
+        Histogram::noop()
+    }
+}