+++ crates/astria-auctioneer/CHANGELOG.md
@@ -16,3 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Use v2 execution API instead of v1 [#2006](https://github.com/astriaorg/astria/pull/2006).
+
+### Fixed
+
+- Fix memory leak in metrics [2221](https://github.com/astriaorg/astria/pull/2221).