+++ crates/astria-composer/CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fetch and submit transactions already in the rollup transaction pool before submitting
     those coming from `eth_subscribe` [#2086](https://github.com/astriaorg/astria/pull/2086).
 
+### Fixed
+
+- Fix memory leak in metrics [2221](https://github.com/astriaorg/astria/pull/2221).
+
 ## [1.0.1] - 2025-03-06
 
 ### Changed